Remove external script (integrate with service)
Let's try something new: the service does not rely on a call to an
external shell script anymore. This is achieved by piggy-backing
the required code within the service file itself.

This is a trial balloon to see if I'm going to utilize this
technique more often in the future. (I'd really prefer my custom
systemd services to be as standalone as possible.)